% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion_data.R
\docType{data}
\name{conversion}
\alias{conversion}
\title{Uniprot PTM names with standarized names and PTM molecular weight effects}
\format{CSV file with standardized PTM names and effect of the molecular weights of the PTM.}
\usage{
data(conversion)
}
\description{
This data was assembled of the author of this package from various sources. The Old ptm names come from the Uniprot PTM names. 
The new PTM names are more computer friendly and merge PTMs that are essentially the same
}
\examples{
data(conversion)


}
\keyword{datasets}
