% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countPTMs.R
\name{countPTMs}
\alias{countPTMs}
\title{A function that takes a protein dataframe and ptm data frame.}
\usage{
countPTMs(proteindf, ptmdf)
}
\arguments{
\item{proteindf}{Data frame of protein data generated by the fetchGenebank function.}

\item{ptmdf}{Data frame of PTM data generated by the ptm function}
}
\description{
This function will allow you to count the number of PTMs in a protein data frame generated by the fetchGenebank function. It will use the data from the PTM dataframe
created by the PTM function. Using the geneid it will link the data and count the PTMs.
}
\examples{

proteindf <- countPTMs(proteindf, ptmdf)
}
\keyword{PTMS}
\keyword{Proteins,}
