% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetchGenebank.R
\name{fetchGenebank}
\alias{fetchGenebank}
\title{A function that takes a list of UID and returns a genebank files}
\usage{
fetchGenebank(UIDs)
}
\arguments{
\item{UIDs}{List of NCBI UIDs.}
}
\description{
This function will allow you to enter a list of protien/gene names and convert the into NCBI unique IDs. It will return a list. 
The first element will be your UIDs the second element will have the protien/gene names that were not matched to an ID.
}
\examples{

df <- fetchGenebankl(UIDs)
}
\keyword{NCBI,}
\keyword{UIDs,}
\keyword{genebank,}
\keyword{parsing}
