% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGenPept.R
\name{getGenPept}
\alias{getGenPept}
\title{A function that takes a list of UIDs and returns a genPept files}
\usage{
getGenPept(UIDs)
}
\arguments{
\item{UIDss}{List of NCBI UIDss.}
}
\description{
This function will allow you to enter a list of protien/gene names and convert the into NCBI unique IDs. It will return a list. 
The first element will be your UIDss the second element will have the protien/gene names that were not matched to an ID.
}
\examples{

df <- getGenPept(UIDss)
}
\keyword{NCBI,}
\keyword{UIDss,}
\keyword{genPept,}
\keyword{parsing}
