% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getUID.R
\name{getUID}
\alias{getUID}
\title{A function that takes a  protien/genelist and returns the Unique ID number of NCBI}
\usage{
getUID(protien_gene_list,
  baseterm = "[Protein Name] AND Mus musculus[Organism]", db = "protien")
}
\arguments{
\item{protien_gene_list}{List of protiens or genes.}

\item{baseterm}{The term that you would like attached to your protien/gene names to make the search as accurate as possible. I.e. "[Protein Name] AND Mus musculus[Organism]"}

\item{db}{The database of the NCBI that you would like to access. I.e. protien}
}
\description{
This function will allow you to enter a list of protien/gene names and convert the into NCBI unique IDs. It will return a list. 
The first element will be your UIDs the second element will have the protien/gene names that were not matched to an ID.
}
\examples{

uids <- getUID(protien_gene_list, baseterm = "[Protein Name] AND Mus musculus[Organism]", db = "protien" )
}
\keyword{NCBI,}
\keyword{UIDs}
