% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/humanMouse.R
\name{humanMouse}
\alias{humanMouse}
\title{A function that takes a list of human genes and finds the mouse homologs}
\usage{
humanMouse(genelist)
}
\arguments{
\item{genelist}{A list of human genes}
}
\description{
This function will accept a list of human gene names and finds their mouse homologs. This gene list can then be used with the getUID function of this package.
}
\examples{

mousegenelist <- humanMSouse(humangenelist)
}
\keyword{biomart,}
\keyword{genes}
\keyword{genes,}
\keyword{human}
\keyword{mouse}
