% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictedMW.R
\name{predictedMW}
\alias{predictedMW}
\title{A function that takes a protein dataframe produces by the countPTMs function along with the PTM data frame to return predicted molecular weight.}
\usage{
predictedMW(proteindf, ptmdf)
}
\arguments{
\item{proteindf}{Data frame of protein data generated by the fetchGenebank function.}

\item{ptmdf}{Dataframe of PTM data generated by the PTM function.}
}
\description{
This function takes a protein dataframe produces by the countPTMs function along with the PTM data frame to return predicted molecular weight. It uses the conversion data frame
to map the PTMs to the correct molecular weight and to produce a predicted molecular weight based on the experimentally verfied PTM data from Uniprot.
}
\examples{

finaldf <- predictedMW(proteindf, ptmdf)

}
