% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptm.R
\name{ptm}
\alias{ptm}
\title{A function that takes a ptm data frame and a data frame of UIDs and UNIPROT IDs.}
\usage{
ptm(ptmdf, res)
}
\arguments{
\item{ptmdf}{Data frame of PTM data generated by the readPTM function}

\item{res}{Data frame of UNIPROT IDs and UIDs produced by the unip function}
}
\description{
This function creates a PTM dataframe that links geneid from NCBI to the UNIPROT PTM data. This will allow this data to link expermintally 
verified PTM data from UNIPROT to protien data from NCBI.
}
\examples{

ptmdf <- ptm(ptmdf, res)
}
\keyword{PTMS,}
\keyword{UID,}
\keyword{Uniprot}
