% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptmClean.R
\name{ptmClean}
\alias{ptmClean}
\title{A function that takes a data frame of PTM modifications created by the PTM function.}
\usage{
ptmClean(ptmdf)
}
\arguments{
\item{ptmdf}{Dataframe of PTM data generated by the PTM function.}
}
\description{
This function will allow you to enter a list of dataframe of PTM data. It will remove spaces and normalize the names for the PTMs. It will also subset the ptmdf,
so that it only includes PTMs that are in your dataset.
}
\examples{

ptmdf <- ptmClean(ptmdf)
}
\keyword{PTMS,}
\keyword{UNIPROT,}
\keyword{cleaning}
\keyword{data}
\keyword{protiens,}
