% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readPTM.R
\name{readPTM}
\alias{readPTM}
\title{A function that takes an output file from the Uniprot command line tool}
\usage{
readPTM(ptmfile = "output.txt")
}
\arguments{
\item{ptmfile}{Output file from uniprot command line tool}
}
\description{
This function takes an output file from the command line tool built to access uniprot. This tool takes a file generated by the unip function. You must request the data be returned as gff format. Sometimes there is a header
that must be removed before the gff data. This tool can be found here https://github.com/jdrudolph/uniprot.
It will return this as an R dataframe.
}
\examples{
Command line Usage:
uniprot retrieve -f gff uniprot.txt output.txt 

ptmdf <- readPTM(ptmfile = 'output.txt')
}
\keyword{PTMS,}
\keyword{Uniprot}
