% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchpro.R
\name{searchpro}
\alias{searchpro}
\title{A function that searches your output based on MW}
\usage{
searchpro(df, range = 1:max(df$molecularweight_predicted), predicted = TRUE)
}
\arguments{
\item{df}{Data frame following the formatting of data frames produced by the prodata package}

\item{range}{Range of values for the molecular weight that you would like to search for. Default returns whole dataset.}

\item{predicted}{If TRUE (default) it will use the molecular weight including the PTM calculations. If FALSE it will use the molecular weight without PTMs.}
}
\description{
This function will allow you to search your data set based on a range of molecular weights
}
\examples{

filtered <- searchpro(df, range = 1000:2000, predicted = TRUE)
}
\keyword{molecular}
\keyword{search,}
\keyword{weight}
