% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unip.R
\name{unip}
\alias{unip}
\title{A function that takes a list of UID and returns a list of UNIPROT IDs}
\usage{
unip(UIDs)
}
\arguments{
\item{UIDs}{List of NCBI UIDs.}
}
\description{
This function converts UID into UNIPROT IDs for the retrival of PTM data from UNIPROT. It writes this to a file to be used in the comman line UNIPROT tool. 
It also returns a list of UIDs attached to UNIPROT ID. This will be used to build the ptmdf.
}
\examples{

res <- unip(UIDs)

}
\keyword{NCBI,}
\keyword{UIDs,}
\keyword{Uniprot}
