% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.R
\name{viz}
\alias{viz}
\title{A function that takes the dataframe created by the predictedMW() function and launched a shiny app}
\usage{
viz(finaldf)
}
\arguments{
\item{finaldf}{Dataframe created by the predictedMW() function}
}
\description{
This function creates a shiny app of the final data frame created by the prodata package and launched a shiny app.
The shiny app allows the user to filter that databased on their prefred molecularweight plus a tolerance.
}
\examples{

viz(finaldf)

}
\keyword{dashboard,}
\keyword{interaction}
\keyword{shiny,}
