% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keyplayers_data.R
\docType{data}
\name{keyplayers}
\alias{keyplayers}
\title{List of keyplayers for protiens active in a diabetic mouse pancreas}
\format{A list of proteins}
\usage{
data(ketplayers)
}
\description{
List of protiens that are active in a diabetic mouse pancreas. This can be used to test the function of this package.
}
\examples{
data(keyplayers)

ids <- getUID(keyplayers, "[Protein Name] AND Mus musculus[Organism]", "protien" ))

proteindf <- fetchGenebank(ids$UIDs)
res <- unip(id$UIDs)

ptmdf <- readPTM(file = 'output.txt')
ptmdf <- ptm(ptmdf, res)
ptmdf <- ptmClean(ptmdf)

proteindf <- countPTMs(proteindf, ptmdf)
finaldf <- predictedMW(proteindf, ptmdf)




}
\keyword{datasets}
