% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/molecW.R
\name{molecW}
\alias{molecW}
\title{A function that takes a dataframe produced by the fetchGenebank() function and returns PTMS and molecular weight}
\usage{
molecW(df, monoisotopic = FALSE)
}
\arguments{
\item{df}{df produced by the fetchGenebank() function or where the sequence column is named seq.}

\item{monoisotopic}{A logical value TRUE or False indicating if monoistopic weights of amino-acids should be used}
}
\description{
This function takes a dataframe produced by the fetchGenebank() function and returns PTMS and molcular weights via extra columns in the data frame.
I can take other data frames if the sequences columns is named seq
}
\examples{

finaldf <- molecw(df)
}
\keyword{Molecular}
\keyword{PTMs}
\keyword{Weight,}
